+++ app/components/framework/CompactAttributeInput/DateRangeWidgetInput/index.tsx
@@ -1,11 +1,9 @@
 import React, { useCallback } from 'react';
-import {
-    DateRangeInput,
-    DateRangeOutput,
-} from '@the-deep/deep-ui';
+import { DateRangeOutput } from '@the-deep/deep-ui';
 import { isNotDefined } from '@togglecorp/fujs';
 import { Error, getErrorObject, getErrorString } from '@togglecorp/toggle-form';
 
+import DateRangeInputWrapper from '#components/DateRangeInputWrapper';
 import NonFieldError from '#components/NonFieldError';
 import WidgetWrapper from '../WidgetWrapper';
 import { DateRangeWidgetAttribute } from '#types/newEntry';
@@ -78,7 +76,7 @@ function DateRangeWidgetInput<N extends string>(props: Props<N>) {
                     <NonFieldError
                         error={error}
                     />
-                    <DateRangeInput
+                    <DateRangeInputWrapper
                         className={styles.input}
                         name={name}
                         onChange={onChange}
